+++ CHANGELOG.md
@@ -1,28 +1,45 @@
 # Change Log
-All notable changes to this project will be documented in this file.
-This project adheres to [Semantic Versioning] (http://semver.org/).
-For change log format, use [Keep a Changelog] (http://keepachangelog.com/).
+
+All notable changes to this project will be documented in this file. This project adheres
+to [Semantic Versioning] (http://semver.org/). For change log format,
+use [Keep a Changelog] (http://keepachangelog.com/).
+
+## [1.2.0] - 2021-11-25
+
+### Added
+
+- PHP 8.1 compatibility
 
 ## [1.1.0] - 2020-11-05
+
 ### Added
+
 - PHP 8 compatibility in `composer.json`
 
 ### Changed
+
 - psr/http-server-handler required version in `composer.json`
 - Update NPM dependencies
 
 ## [1.0.3] - 2020-11-05
+
 ### Fixed
+
 - Composer provide rule for package `psr/http-server-handler-implementation`
 
 ## [1.0.2] - 2020-09-11
+
 ### Changed
+
 - Display of config in debug console fixed
 
 ## [1.0.1] - 2020-09-02
+
 ### Changed
+
 - Print response by sequence to prevent exhausted memory error
 - Update NPM dependencies
 
 ## [1.0.0] - 2020-05-29
+
 First version